Add information about current runtime for check
When using `resman -c`, you get some info about the potentially running
job. This job might not have a duration (eg. instead running a command),
and then it is hard to have any idea of when to check again for
availability.

This commit outputs one extra line to `resman -c` in case the current
running job does not have a duration: One which shows for how long the
current job has ran so far. If this is several minutes, you don't have
to check back in every few seconds.