Add a red `!` to the prompt when system is reserved
This is done by updating the global `RESMAN_ALERT` every time the prompt
is refreshed. This variable can also be included in custom prompts.

It updates the variable by checking the global lock file. This can be
costly as it has to parse JSON, so it therefore truncates the file when
the previous reservation has ended, making future checks faster.
However, it does introduce some overhead when the server is reserved.
Instead of parsing JSON, you could hard-code the assumption that the
end_time is first in the record, doing it all in bash, but that is less
robust.